Use fabsf instead of fabs for floats

+++ src/drivers/vmount/vmount.cpp
@@ -107,8 +107,8 @@ struct Parameters {
 		       mnt_mode_out != p.mnt_mode_out ||
 		       mnt_mav_sysid != p.mnt_mav_sysid ||
 		       mnt_mav_compid != p.mnt_mav_compid ||
-		       fabs(mnt_ob_lock_mode - p.mnt_ob_lock_mode) > 1e-6 ||
-		       fabs(mnt_ob_norm_mode - p.mnt_ob_norm_mode) > 1e-6 ||
+		       fabsf(mnt_ob_lock_mode - p.mnt_ob_lock_mode) > 1e-6f ||
+		       fabsf(mnt_ob_norm_mode - p.mnt_ob_norm_mode) > 1e-6f ||
 		       mnt_man_pitch != p.mnt_man_pitch ||
 		       mnt_man_roll != p.mnt_man_roll ||
 		       mnt_man_yaw != p.mnt_man_yaw ||
